Add Thrift source information to type_system.thrift
Summary:
# Why do "names" belong in `type_system.thrift`?

Names are very "core" to Thrift. The fact that `.thrift` files exist, and that each Thrift file has unique mapping of names to definitions is something that is that we *never* envision will change in the evolution of Thrift. In practice, this means that users of Thrift **vastly** prefer using names, not URIs wherever possible.

We've already seen some cases where such "metadata" would be handy:
1. TXL prototype has build `ScopedTypeSystem`.
2. Pretty printing of `type_system.thrift` into a set of "equivalent" `.thrift` files.
3. Debugging / error messages.

# Why the tuple `(<locator>, <name>)`?

Most type systems will derive from a `.thrift` file *somewhere*. This representation is the natural format to represent the origin of those types.

The locator is a URI so that, within Meta, we can set up conventions around the semantics of those URIs (e.g. URI could point to a snapshot stored within a schema store).

# What about non-"name" metadata?

Other IDL concepts are *not* relevant to the type system as described by the Object Model. For example, `exception`s in the IDL are not captured as part of the symbol table information. They may belong in another spec derived from the Object Model (something like an RPC model).

We *may* introduce non-name metadata in the future. We can do so in a backwards-compatible manner by adding a 3rd field to `SerializableThriftSourceInfo`.

Reviewed By: aristidisp

Differential Revision: D75806744

fbshipit-source-id: b8ed2375577b9a2199868a7537a2ba21bd5600bc