Make serialize builtin use lazy class flavors for KindOf[Lazy]Class
Summary:
Deserialization support for lazy classes was added in {D25111001} to support C::class const values in HackC. This changes the `serialize` native function to emit `l` flavors when it gets a class or a lazy class. Other kinds of serialization will remain as they are for now.

Removes the class-to-string helper calls since we no longer need to log those.

Reviewed By: ricklavoie

Differential Revision: D76548453

fbshipit-source-id: 884f33b4452bffc9aa782471e6fc1d2071f645d0