Fix test flakiness for redeclaration daemon
Summary:
## What

Make tests less flaky. This is a no-op from the user point of view

## Background

In notebooks, we rewrite code before sending to HHVM.
([wiki](https://www.internalfb.com/wiki/Hack_in_Bento_notebooks/Hack_Notebooks_-_Internal/)).

One of the ways we rewrite code before sending to HHVM is to print the last line, which we do differently depending on whether HHVM is stopped at a breakpoint: this is because only `debugger_dump` will actually print anything if the user is stopped at a breakpoint.

At a breakpoint, we use `debugger_dump` which prints some pretty raw, old-school stuff:
 {F1979193233}

Not at a breakpoint, we use `BentoHack::display`, which pretty-prints where it can:

 {F1979193232}

We use presence of a particular file (shared mutable state) to detect if HHVM is paused.
Shared mutable state is not great for test reliability.

## Hypothesis

Our language server integration tests create the file that indicates that HHVM is stopped at a breakpoint.
Perhaps our language server tests were running at the same time as the redeclaration daemon tests, and causing the redeclaration daemon code (while under test) to think that HHVM was stopped at a breakpoint and change its behavior.

## The fix

Remove the shared mutable state in tests.

We tell the redeclaration deamon where to look for the magic file that indicates HHVM is paused:
- in most tests, we say not to look for the file, so we can test the normal code path–the behavior where HHVM not being paused
- in tests specifically for the pause behavior, we specify the path `/dev/null`, which is pretty much guaranteed to exist or something else is very wrong.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D76593123

fbshipit-source-id: 3163bd47846b7cf62f9372de0cc92172f425ed9a