Export generic type var names
Summary:
This diff introduces:
1. a struct TParamInfo as part of func Body
2. a [ClassName] to store template param names in Class (change to TParamInfo in future diffs)

Subsequent diffs will merge all template param information including upper bounds and reified generics into this struct.

These fields are threaded through various classes in the runtime and eventually stored in the GenericsInfo struct in func and class which will be exposed via reflection.

It's worth mentioning that I chose to pass the type var names separately from the reified user attribute because the presence of the reified attribute is currently used to indicate whether the func or class has reified params. I didn't want to overload this attribute to also have non-reified params. Given that I plan to consolidate the reified user attribute with the newly introduced struct, I thought it would be best to leave this undisturbed.

Reviewed By: jano

Differential Revision: D75026944

fbshipit-source-id: b874155a699da66ef67c0ff703a4a272416fc57e