Make APC and internal serialize use lazy classes for class pointers
Summary:
The logic for APC serialization is inconsistent for lazy classes and class pointers depending on the container. Before this diff, `apc_store` on a bare class pointer works fine, but in a `Pair` it comes back as a string.

With this diff, I'm still inconsistent between storing the pointer (see `APCTypedValue(const Class*)`) vs. a lazy class, but it eliminates the stringy-ness. The unserialize logic for `c` kinds appears to only have been added for the record-replay logic, so I would like to test it separately. Filed T227444162 to track unification.

 ---

Code pointers:

https://www.internalfb.com/code/fbsource/[a3b0b69363b85c03111e37c07a59651631b2d9ca]/fbcode/hphp/runtime/base/apc-handle.cpp?lines=89-101

https://www.internalfb.com/code/fbsource/[a3b0b69363b85c03111e37c07a59651631b2d9ca]/fbcode/hphp/runtime/base/apc-collection.cpp?lines=88-94
https://www.internalfb.com/code/fbsource/[a3b0b69363b85c03111e37c07a59651631b2d9ca]/fbcode/hphp/runtime/base/collections.cpp?lines=131-146

https://www.internalfb.com/code/fbsource/[a36f581b09a9239b7f3ba740615a512e7a8255e9]/fbcode/hphp/runtime/base/variable-unserializer.cpp?lines=1217-1223

Reviewed By: jano

Differential Revision: D76558975

fbshipit-source-id: 5f631c32198b4c4b4227a13c8d82e7be066bb6e0