Write jit profile for debugging when not jumpstarting
Summary: When we get a crash on a host that doesn't jumpstart it would still be valuable to be able to extract a JIT profile to reconstruct the crashing translations. This diff adds a new runtime option (Eval.JitSerializeDebugLocation) to indicate where this extra profile should be written, and changes the serializer to also map in any serialized profiles when not in deserializing mode.

Reviewed By: jano

Differential Revision: D76632659

fbshipit-source-id: 8a70295e999fdadde75b753e1889291eec9bb974