Update the notion of `this` when accessing type constants through intersections and unions
Summary:
When accessing a type constant through union and intersection types, the notion of `this` contained in the `expand_env` should be updated for each element.

This is particularly important when the access type constant contains a fixpoint refinement and one of the elements of the union/intersection is an expression dependent type. Unless we update `this` to that expression dependent type we will fail to properly expand the type constant.

Reviewed By: andrewjkennedy

Differential Revision: D76598028

fbshipit-source-id: 24218ddcfeb42bc00aec50516327e6e926db0e18