Expand function pointer check, unify with special_function emitter, refactor a bit
Summary:
The emitter has a set of "special" functions where for calls, it emits special bytecode sequences. A bunch of these do not have runtime definitions, although with a few notable exceptions like `array_mark_legacy`. So, requesting a function pointer from them is an error. There was already a check for UNSAFE_CAST; expand it to the rest of these that are defined in HHI (we can skip over the __Systemlib ones).

kills the unused EmitterSpecialFunctions module and renames the old SpecialFunctions so we can share the name.

For future improvement we can
- kill PreNamespacedFunctions module and just inline the echo checks
- when namespacing is unified, push this check down to the parser so you get a compiler error instead of a runtime error at resolution

Reviewed By: viratyosin

Differential Revision: D75290177

fbshipit-source-id: a764285d8b690baa6dc4af7ce2ea11dd1a636dfa