Use OR to list out possible source types
Summary:
This allows us to parse all the type information as single argument as opposed to a variadic list. Two benefits:

This makes it less sketchy to add source-specific flags (see comments in the next diff).
This makes it more clear what the comma separated types are actually supposed to mean.

Reviewed By: jano

Differential Revision: D76579398

fbshipit-source-id: 486c58a00d0f2385fb304ec2a76b0e0e419b17c5