introduce FilesToIgnore.watch_spec
Summary:
This diff introduces `FilesToIgnore.watch_spec` as a specification of what files we want Watchman or `Edenfs_watcher` to watch for changes. It's just a list of extensions and file names that we are interested in.
The idea is to specify what files we want to watch once, and then have both Watchman and `Edenfs_watcher` consume that specifcation.

For Watchman, `FilesToIgnore.watchman_server_expression_terms` is a JSON object that encodes the "expression term" that we send to Watchman to tell it what files we want to watch.
This diff changes this situation such that the source-of-truth regarding what files we want Watchman to watch for is now `server_watch_spec`, of type `watch_spec`.
We then generate `watchman_server_expression_terms` from `server_watch_spec`.

In a subsequent diff, `Edenfs_watcher` will consume `server_watch_spec` directly to configure it.

Differential Revision: D76749014

fbshipit-source-id: 4e9d923c6bc24a8a34c78e26a72d7a01ed446bb8