allow running hh_server in foreground in tests
Summary:
This is a quality-of-life improvement for `test_edenfs_file_watcher.py`:

Normally, these tests start `hh_server` with `--daemon`, which causes its logging to go to dedicated files. This diff allows setting the environment variable `DEBUG_EDENFS_WATCHER_TEST_HH_SERVER_FOREGROUND`, in which case we omit `--daemon`. Instead we redirect its output straight into the test process.

As a result, we get the server's logging interleaved with anything else happening in the test (e.g., the Eden logs, calls to external programs). I'm using this when debugging to see more easily what's happening when.

Differential Revision: D76747500

fbshipit-source-id: ea056129734ee33601dd374473725ef9fc116f37