fix server crash on folder rename
Summary:
This fixes a bug introduced in D76124936:
If we run `Edenfs_watcher` on a folder that's not the Eden mount point, we called `glob_files` in the wrong way, causing Eden to throw an error and making the server exit.

What's worse is that this bug was not discovered by the test (`test_folder_rename` run by `EdenfsWatcherNonMountPointRepoTests`) for this exact scenario. The test passed because it renames a folder containing files with type errors, and then runs `hh check` to verify that the error messages show the updated paths (to check that the server has indeed noticed that the files inside the folder were moved).

This test passed because when the server crashed, it was restarted by the monitor, after which it would re-check everything and pick up the moved files this way. To avoid this problem in the future, I've added a check to all `Edenfs_watcher` tests to make sure that the server hasn't restarted.

Differential Revision: D76582277

fbshipit-source-id: 1fec471ef466f108e22ee87f93f312c80ea8a69e