Fix time widening bug in JIT
Summary:
When intersecting RAT Array types we choose a specialization arbitrarily to avoid a potential O(n) traversal. While this is usually fine, there are cases where widening a type can cause things to break.

https://www.internalfb.com/code/fbsource/[3f4279bf78a40f1f64fbc06ea1efd402de30943d]/fbcode/hphp/runtime/vm/jit/type-specialization.cpp?lines=28-30

In particular in a crash observed in HHVM trunk widening a `Vec=Vanilla:N([Str])` type to `Vec=Vanilla:N(InitCell)` caused us to lose type information on values loaded from the Vec. As this widening was able to occur as a result of reflowing types following a refineTmps pass we had already performed specializations on the element loaded from the vec that depended on its type being a string.

```
Error: failed type check on operand 0
   instruction: (373) t134:Str = ConcatStrStr t114:InitCell, "/" -> B103<Catch>
   was expecting: Str
   received: InitCell
   from: (310) t114:InitCell = LdPtrIterVal<InitCell> t294:Vec=Vanilla, t113:PtrToElem
```

Reviewed By: jano

Differential Revision: D76878556

fbshipit-source-id: d21d7c9119a49b5888d697eb89b1ba8b2fa39c66