Do not require parameters for Memoize in hhi files
Summary:
There are some hhi files we generate based on their implementation in the HSL that is shipped with HHVM/Hack. If these functions are marked as `<<__Memoize>>` then we also mark them as `<<__Memoize>>` in the .hhi we generate. This creates an issue if the `disallow_non_annotated_memoize` config is enabled, because then when we type check these .hhi files we report an error.

This is a problem because if there is an error outside of the root directory, an [invariant_violation](https://www.internalfb.com/code/fbsource/[d37fba14af8f]/fbcode/hphp/hack/src/client_and_server/server_progress.ml?lines=501-505) is triggered which crashes the server on start up.

The fix I propose for now is to not report this error within .hhi files. This should be safe given that <<__Memoize>> is an implementation detail and doesn't change the interface/API for a function.

Reviewed By: KendallHopkins

Differential Revision: D76938619

fbshipit-source-id: 63e705f18f8e80da54765b5e03447c92ef15955c