Fix bug which disallows all wildcard hints in poly lambda
Summary:
Fixes a bug which raises a naming error if you use a wildcard hint anywhere in a polymorphic lambda. The intended behavior is that they should be disallowed in parameter or return position but they should be allowed in the function body e.g. in `is` refinements.

The diff updates the corresponding validation pass to modify the context when we enter the function body.

Reviewed By: mheiber, madgen

Differential Revision: D77018041

fbshipit-source-id: dfc556d96ec8304bb01f525fcabbedd8a1716d29