Warn on calling static methods on traits
Summary:
Calling methods on traits is a footgun in the language.

The footguns stem from the fact that traits are incomplete fragments of a class but are distinct from the classes they get used. So calling private method on `MyTrait::foo()` inside `MyTrait` is different than calling `self::foo()` and similarly static properties of `MyTrait` are different than classes that include the trait.

With this diff, we produce on all direct static calls on traits. The intention is to promote it to a Hack error later on.

While at it, I added a reason to the require class version of the same check.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D77024038

fbshipit-source-id: 4f16088a5f304a049662c4ecbf44ba423fce194c