let the kernel trap the process after some fatal signals
Summary:
When the `SEGV` signal handler runs and dumps the stack, it then raises `SEGV` again for the next signal handler or the default action. But this may lose information by the time the signal makes it to the debugger or the coredump.

Instead, let the `SEGV` signal handler return after removing itself, let the thread run the same instruction again raising `SEGV` again, and this time the default action would let the kernel trap the signal with full information.

Reviewed By: Jlalond

Differential Revision: D77041890

fbshipit-source-id: 5858edce42de65d8011203b9ecd9acacb4d002c2