Kill empty pairs of EnterInlineFrame / LeaveInlineFrame
Summary:
~9% of all EnterInlineFrame instructions are immediately followed by
LeaveInlineFrame. Plus there are more pairs that wrap empty pairs. This happens
when an inlined stack of methods simplifies to a constant.

Extend the initial pass that walks instructions and marks them as live to keep
track of open EnterInlineFrames and do not mark empty pairs as live.

In some cases this allows reduction of the whole block to an empty block,
possibly allowing killing a conditional jump if the other branch is equivalent.

Reviewed By: ricklavoie

Differential Revision: D77059562

fbshipit-source-id: e44cbad69b4e2e48a61ef5b3189989ada26e188c