Ban accessing static properties on traits
Summary:
Accessing static properties off of a trait is a footgun.

When one `use`s a trait the using class also has a static property of the same name, but this static property is completely different than the one that belongs to the trait.

Consequently, updating the property on the trait does not reflect any changes on the static property of the class and vice versa.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D77089496

fbshipit-source-id: 33158f16d2e12643e7fd0584d499a6700d709acc