Support reuse of same client/server in LocalThriftConnection
Summary:
Server read/write method expect that the Transport buffer are already empty. In production, this is ensured as we create a new instance of Server along with new buffer.

When used with multiple requests, `LocalThriftConnection` reuses the same transport  for client/server read/write. If one request doesn't use the entire buffer, the next request would fail. Clear the buffers before staring a new request.

Differential Revision: D76958354

fbshipit-source-id: 259177eeefee9f6a2dea9dbb08646e59acc10248