Remove poll requirement for messages after hphpd starts
Summary:
## Context
Fbdbg is an unowned and unmaintained Hack debugger that's used in Hack Bento Notebooks and on the command line as a repl.

Although fbdbg enables debugging capabilities, it's rarely used among Hack programmers, with the VSCode debugging extension being preferred.

Deprecating fbdbg allows us to
1. Reduce support burden on the HHVM team
2. Enable more capabilities for Hack notebooks easily

["Deprecating fbdbg" Google Doc](https://docs.google.com/document/d/12CyH78_LT6CbXMQF603-_75DwkVlpmYo2MW9bfEFMM4/edit?tab=t.0)

## This diff
Right now, the hphpd client only starts reading messages when informed by the poll. This doesn't work for streaming messages since it's possible that there are multiple messages in the socket when the client reads from it. It should only wait for timeouts when streaming is completed.

The bijection is also not necessarily correct, since a [separate thread in the hphpd server](https://fburl.com/code/kuzrx6x0) will send occasional `CmdSignal` pings to the client while other threads are running.

Future diffs:
* The 1 second poll timeout makes hphpd seem slow when messages are in the buffer and not the socket, since there's a 1s/message delay when messages are just waiting in the read buffer. We should reduce the timeout and add some logic to avoid the poll when the read buffer contains unread messages.
* Integration into the Hack Bento kernel
* Add JIT to hphpd

## This stack
Currently, streaming doesn't work without a sleep call since the hphpd client doesn't expect messages to arrive so quickly. At the moment, the hphpd client assumes that if poll(fds) says the file/socket is ready, exactly only one message is in the buffer, so
1) It[ resets the buffer after reading one message](https://fburl.com/code/2l41j45t), so it ignores all of the latest messages read from the socket earlier. Fixed in D76405423.
2) It only uses one buffer for reading and writing, without a lock.  When there are multiple messages read from the socket, the client can't respond to the server's pings, causing the server to kill hphpd. Fixed in D76405421.
3) The hphpd client[ expects only one message per poll](https://fburl.com/code/b2bqwp3x), so it stalls out when there's messages in the buffer that have already been read from the socket (and no more messages are expected). Fixed in D76405422. there is no other time to reset the buffer when reading.

Reviewed By: nt591

Differential Revision: D76405422

fbshipit-source-id: e657720c5e09e12097a45caa1bbfdb8b7ccea11a