Don't reset the buffer for streaming messages
Summary:
## Context
Fbdbg is an unowned and unmaintained Hack debugger that's used in Hack Bento Notebooks and on the command line as a repl.

Although fbdbg enables debugging capabilities, it's rarely used among Hack programmers, with the VSCode debugging extension being preferred.

Deprecating fbdbg allows us to
1. Reduce support burden on the HHVM team
2. Enable more capabilities for Hack notebooks easily

["Deprecating fbdbg" Google Doc](https://docs.google.com/document/d/12CyH78_LT6CbXMQF603-_75DwkVlpmYo2MW9bfEFMM4/edit?tab=t.0)

## This diff
This diff adds the ability to not reset the buffer for streaming messages. The client will continue to reset the buffer for other types of cmd messages; since the thrift buffer isn't a ring buffer, there is no other time to reset the buffer when reading.
## This stack
Currently, streaming doesn't work without a sleep call since the hphpd client doesn't expect messages to arrive so quickly. At the moment, the hphpd client assumes that if poll(fds) says the file/socket is ready, exactly only one message is in the buffer, so
1) It[ resets the buffer after reading one message](https://fburl.com/code/2l41j45t), so it ignores all of the latest messages read from the socket earlier. Fixed in D76405423.
2) It only uses one buffer for reading and writing, without a lock.  When there are multiple messages read from the socket, the client can't respond to the server's pings, causing the server to kill hphpd. Fixed in D76405421.
3) The hphpd client[ expects only one message per poll](https://fburl.com/code/b2bqwp3x), so it stalls out when there's messages in the buffer that have already been read from the socket (and no more messages are expected). Fixed in D76405422.

Reviewed By: nt591

Differential Revision: D76405423

fbshipit-source-id: 9631ff2b85b7acc009c1206c525f55f1356c89ba