Use the JIT in notebook mode
Summary:
## Context
Fbdbg is an unowned and unmaintained Hack debugger that's used in Hack Bento Notebooks and on the command line as a repl.

Although fbdbg enables debugging capabilities, it's rarely used among Hack programmers, with the VSCode debugging extension being preferred.

Deprecating fbdbg allows us to
1. Reduce support burden on the HHVM team
2. Enable more capabilities for Hack notebooks easily

["Deprecating fbdbg" Google Doc](https://docs.google.com/document/d/12CyH78_LT6CbXMQF603-_75DwkVlpmYo2MW9bfEFMM4/edit?tab=t.0)

## This diff
The JIT is [important in improving the performance](https://fb.workplace.com/notes/1369517753981647) of the debugger and Bento notebook. While it's currently enabled by default on fbdbg, it should also be enabled for hphpd. This diff enables the JIT for hphpd notebook mode. Eventually, hphpd should use the JIT by default.

NOTE: Historically, the debuggers did not use the JIT since it did not work well with breakpoints and debugging. These issues should be fixed now that fbdbg uses the JIT, and either way, we will not be using hphpd for debugging.

Future diffs:
* The 1 second poll timeout makes hphpd seem slow when messages are in the buffer and not the socket, since there's a 1s/message delay when messages are just waiting in the read buffer. We should reduce the timeout and add some logic to avoid the poll when the read buffer contains unread messages.
* Integration into the Hack Bento kernel

Reviewed By: nt591

Differential Revision: D76555555

fbshipit-source-id: 2ec42a10c863f6680b9ef6ce31ad68865bd31ee7