codemod for adding __NeedsConcrete
Summary:
This is the script what I've been using for getting stats on Safe Abstract, see fb-only section.

It runs `hh --re` with `--needs_concrete=true` in a loop, fixing errors as it finds them until it reaches a fixed point.

This is not a work of great beauty and performance, but it gets the job done.

## Follow-on work

- power dashboards

Reviewed By: andrewjkennedy

Differential Revision: D76904286

fbshipit-source-id: f2386edb908d02644ce1c3d1e54deb3ee75ebf7b