make test_hhconfig_change more robust
Summary:
I've noticed that `test_hhconfig_change` in `test_edenfs_file_watcher.py` can be flaky.

The test checks that we restart the server after `.hhconfig` was changed.

Currently, it modifies that file, runs `hh check` (with the intention to wait wait until the server restarted), and checks the server log for a message indicating that we restarted due to an `.hhconfig` change.

The problem is that if we run `hh check` such that it connects to the old server, and stays connected to it until the point when the server restarts, `hh check` will exit with an error message stating that the server disconnected unexpectedly, making the test fail.

This diff changes the test such that we create a file that has a type error according to the old `.hhconfig`, and the `.hhconfig` change makes the file well-typed. This means that after the `.hhconfig` change, we can just loop until we get an error code of 0 from `hh check`.

Differential Revision: D77027105

fbshipit-source-id: 46a7bbf175b12a66c8ccb600b51d0fb13f7e8ba1