fix bug in notification pipe consistency checks
Summary:
This fixes a bug in a consistency check we have in place for the notification mechanism `Edenfs_watcher` provides to its clients.

We have a consistency check in place to make sure that the contents of the notification pipe and the contents of the shared `Changes` queue do not get out of sync: We write objects of type `NotificationPipeEntry` to the pipe, which is a simplified version of the journal position at which we added the corresponding entry to the `Changes` queue.
When reading entries from the shared queue, we then read the same number of `NotificationPipeEntry` values from the notification pipe. We then check that the final `NotificationPipeEntry` value  matches `last_journal_pos` in the shared state.

This consistency check is no longer correct as of `D76755006`: We may now receive updates from `Eden` and update `last_journal_pos` in the shared state without adding an entry to the `Changes` queue, because the updates from Eden were filtered out (they affected files we are not interested in).

To fix this, I've updated the shared queue so that it now compares values of type `(Changes, NotificationPipeEntry`) instead of just `Changes`. In the new design, every queue entry itself tells us what the corresponding `NotificationPipeEntry` value in the notification pipe should be like. `last_journal_pos` is no longer involved in the consistency checks.

Differential Revision: D77148089

fbshipit-source-id: d7c3c13b6826ffe01b90fa05471327712af02fca