Make collectAll take tasks by value
Summary:
This makes it much easier to have `collectAll()` properly support immediately-awaitable tasks (`NowTask` for now, eventually `SafeTask` too).

Since `collectAllImpl()` was already taking awaitables by-value, so this change should be a safe no-op.

 ---

General note on immediately-awaitable task support in `folly/coro` --

The only awaitable that I'm aware of that cannot be taken by value is `Baton`, and ... it's not very likely that somebody is trying to use `collectAll` with `Baton`.

I'm fairly sure that the right long-term resolution is for `Baton` to offer a reference-wrapper awaitable that **can** be taken by-value. Then, we can do away with all the `must_await_immediately_v` bifurcations in `folly/coro`.

Reviewed By: ispeters

Differential Revision: D77265296

fbshipit-source-id: 1b465b00c79b91f824a4021ff991ed64cb5d6a03