Remove nullType
Summary:
In the discussion for the proposal to add support for the coalesce operator it's been decided that the definition of a custom null type in expression tree DSLs is unnecessary, so in this diff I'm removing all references to `nullType`.

The proposed coalesce operator implementation uses the `??` operator in the virtual expression, as opposed to the virtual expression of other binary operators that is implemented as a call to instance methods such as `__ampamp`, `__barbar`, etc.

This decision was driven by the special way null values are handled in Hack and by the fact that it would have been confusing to have the ability to define a custom null value that wouldn't have benefited from type refinement in the virtual expression using `??`.

Reviewed By: dlreeves

Differential Revision: D77225537

fbshipit-source-id: 3874c8d479ae16313eab91817891ecb26e8cc0cb