Plug in a function to publish metrics to scuba
Summary: Plugs in the log_metrics_tag pluggable function with an implementation to export metrics to scuba. There is a complexity that the standard loggers have a build dependency on the thrift compiler, meaning the thrift compiler cannot depend on the logger and create a circular dependency. Instead we use the `scribe_cat` cli tool to export the metrics

Reviewed By: iahs

Differential Revision: D76854422

fbshipit-source-id: e2547240c977b43c1787e8ea3b27bc8e6c2ba87c