Warn against static property redeclaration
Summary:
Static property redeclaration is a footgun because it causes a separate static property be generated. In that sense it is not a (re)declaration but just a separate declaration.

```
<?hh

class C {
  public static vec<int> $p = vec[];
}

class D extends C {
  public static vec<int> $p = vec[];
}

<<__EntryPoint>>
function main(): void {
  C::$p[] = 1;
  D::$p[] = 2;
  var_dump(C::$p); // vec[1]
  var_dump(D::$p); // vec[2]
}
```

This diff warns against redeclaration.

We'd like this to be an error eventually.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D77366104

fbshipit-source-id: ad28a6c1da574af4afd5e268bf17922e396b6aaa