ServerNotifier.init fallback logic
Summary:
This diff implements logic so that if `edenfs_file_watcher_enabled` is set, but we fail to initialize an Eden-backed `ServerNotifier.t`, we fall back to the next enabled option (which will usually be Watchman).

This diff also introduces a dedicated event `EDENFS_WATCHER_FALLBACK` that we log to `hh_server_events` whenever that happens.

Differential Revision: D77296408

fbshipit-source-id: 2db84fc8755d74ce33243eeded637da58b89f0ae