workaround to defer file changes while Meerkat is running
Summary:
When using Watchman for file watching, we set it up in way so that any file changes that happen while Meerkat is running are deferred until when it's finished.

The Eden file change API will support such a deferral mechanism eventually, but doesn't do so today. In the meantime, this diff implements a workaround:
- The new module `MeerkatTracker` is mostly a stripped down version of `ServerRevisionTracker`, but for the `meerkat-build` state. The idea is that `ServerNotifier` tells it about any state change, and the module keeps track of the `enter` and `leave` events for the `meerkat-build` state. This way, we can efficiently answer the question "is Meerkat currently running?"
- I'm intentionally re-using the logic from `ServerRevisionTracker` unchanged to handle state transitions, as it implements some logic to smoothen things (e.g., handle out-of-order or missing events).
- In `ServerNotifier.get_changes_{sync,async}`, while `MeerkatTracker.is_meerkat_running` is true, we continue processing all state transition events, but hold back any actual file change events.

The only controversial part about this change may be the following: Since the deferral is happening entirely within `SeverNotifier`, `Edenfs_watcher` is unaware of the deferral and will push notficiations to the notification file descriptor. As a result, I had to insert a 250ms delay into `ServerNotifier.get_changes_{sync,async}` *but only if we are deferred anyway*. See the comments there for details.
That's ugly, but doing something more sophisticated would involve adding the deferral mechanism into `Edenfs_watcher` itself. Since the work in this diff is a workaround while we wait for the corresponding Eden feature to land soon, I would like to avoid that.

Differential Revision: D77305320

fbshipit-source-id: 7674530923b22cd930d2a7cfed9cf05a2b5ad4ba