simplify Tgeneric localization
Summary:
This is a refactoring that results from consolidating code based on the fact that we don't encounter higher-kinded types anymore.

This diff just simplifies the localization code for `Tgeneric`: We can throw away the helper function `localize_tgeneric` it uses, based on the follow observations:
- `replace_with` is always `None`
- `arg_kinds` is always `[]` (since there are no more HK types)

Using this information, we observe that the function always returns `((env, None, []), mk (r, Tgeneric x))` and can inline it.

Reviewed By: enetsee

Differential Revision: D77514432

fbshipit-source-id: 71ca9fbae464c3a104b90da58ff06e4e5a733550