remove type arguments from Habstr
Summary:
This is the counterpart to D73855688 (removing type arguments from `Tgeneric`) for type hints: It removes type arguments from `Habstr`, the hint-level representation of type parameters.

The only non-trivial change is how we report errors when we see something like `T<int>`. With HKTs gone, this is always disallowed. The corresponding error was previously produced in `Naming_elab_hkt`, but that file is going away soon.
Instead, the same error is now emitted in `Naming_elab_happly_hint` at the point where we would previously create a `Habstr ("T", l)` with a non-empty list `l`.

Reviewed By: enetsee

Differential Revision: D77515771

fbshipit-source-id: e19d36a1cd9296560ff5e2b7a18731390d234b95