update Typing_type_integrity
Summary:
This is a refactoring the preservers behaviour. It updates function names inside `Typing_type_integrity` and flags used elsewhere to decide if we should call functions in that module.

## Background on  `Typing_type_integrity`

Years ago, I created this module under a different name, to perform well-kindedness checks, as part of the HKT implementation.
Since then, more checks on types have been added to the module (performing accesibility checks), and it has gotten the name it has today: `Typing_type_integrity`.

However, the names of the functions in the module still reflect the old purpose (e.g., `check_well_kinded_hint`).

## In this diff

Now that the HKT implementation is going away, the original purpose of the module (i.e., well-kindedness checking) is greatly reduced: All it still needs to do is to report `Type_arity_mismatch` errors if you provide the wrong number of type arguments somewhere.

Therefore, I'm renaming the functions in there to mention "integrity checking", in line with the module name. I'm renaming some flags in other `Typing*` modules accordingly.

A better name to describe what the module is doing instead of  "type integrity checking" would probably be "type wellformedness checking". But we already have the separate module `Typing_type_wellformedness`! The latter uses `Typing_type_integrity` so it may be a good idea to consolidate them, but I'm not planning to do that as part of the HKT removal.

Reviewed By: enetsee

Differential Revision: D77515427

fbshipit-source-id: be2cd0d0e1894474924ef0cc63ff4f4fafecb953