remove Typing_kinding_defs.Simple
Summary:
This is a refactoring the preservers behaviour.

## Types in Typing_kinding_defs

The HKT implementation used two notions of kinds:
1. "Full" kinds, represented by `Typing_kinding_defs.kind`. Full kinds basically mean "give me all the information we have about a type parameter in the environment". In particular, they carry bounds. Originally, they could be higher-order in the sense that a `kind` carried a list of parameters itself of type `kind`, but those were already removed earlier as part of the HKT removal.
2. "Simple" kinds, represented by `Typing_kinding_defs.Simple.kind`. The HKT implementation never supported higher-kinded types with bounds. All well-kindedness checks were performed on simple kinds instead, obtained from full kinds by effectively throwing away the bounds. Simple kinds still carried around some limited information about bounds, but that's not relevant for this diff.
It's best to think of simple kinds to just be the kinds you would built from constructors `*` and  ` -> ` in the literature.

## In this diff

This diff simply deletes the notion of simple kinds, by deleting the `Simple` submodule of `Typing_kinding_defs`. They we only needed for well-kindedness checks necessary in the presence of higher-kinded types.

The "full" kinds stay: They represent the information we carry in the typing environment about type parameters. We could rename `Typing_kinding_defs.kind` to something like `tparam_info`, but using the term "kind" for that seems fine.

This diff also ...
- deletes some code elsewhere that becomes dead due to deleting the `Simple` module.
- makes some changes to `Typing_type_integrity`, so that it uses values of type `decl_tparam` instead of `Typing_kinding_defs.Simple.kind`. This module is overhauled in the next diff in this stack anyway.

Reviewed By: enetsee

Differential Revision: D77514969

fbshipit-source-id: ac686780e5bafade1f29a01796b6ff73175a5fbd