Fix conversion from floating point to string
Summary:
We discovered a conversion bug in the Squangle code when converting a floating point value to a string.  The wrong type was checked and the result was an empty string being returned, losing the data.  If the code then attempted to convert back to a floating point value we would get a folly::ConversionError since the input was empty (S535577).

The fix is to check the right types (there are five types we need to check - `bool`, `int64_t`, `uint64_t`, `double`, and `folly::StringPiece`).  In addition I added a numeric conversion test.  This test fails for the Thrift protocol before the fix is applied.

Differential Revision: D77469879

fbshipit-source-id: dfdf4060e4039a6167690f341f2e3a0b3d314fd0