Add support for StopTLS v1 and v2 in the stress test
Summary:
## Context

The high-level intention of this stack of diffs is to implement and integrate a custom TLS record layer, called StopTLSV2RecordLayer, which supports selective encryption modes for Thrift connections. The goal is to enable flexible encryption policies, where certain records can be sent in plaintext, while others are encrypted, depending on the specific requirements of the application.

## Key changes:

1. Implemented support for StopTLS v1 and v2 in the stress test.
2. Resolved a stuck issue by introducing a temporary EventBase (EB) to drive stopTLS v1 negotiation, thereby preventing conflicts with the stress test's EB and ensuring seamless execution.
3. Created test cases `RequestResponseTm_32k` and `RequestResponseTm_32k_unencrypted` for StopTLS v1 and v2 testing.
4. Fixed a dev-mode crash issue caused by releasing `warmupDoneTimeout_` and `testDoneTimeout_` in a  none-EB thread.

## Summary

This diff introduces changes to support stress testing for stopTLSv2 in the Thrift server and client code.

Reviewed By: robertroeser

Differential Revision: D77256423

fbshipit-source-id: afec208cdd7162ed8b461db789a614e39154a606