use F14FastMap
Summary:
F14FastMap is a drop-in replacement for `std::unordered_map` where iterator stability is not required. This describes our use case, where we only access the map after all values are populated.

While this doesn't make any change in the benchmark, there's a reasonable chance it will matter for larger structs.

Reviewed By: iahs

Differential Revision: D77688418

fbshipit-source-id: e4e5a082869f9ec09ccbcec59436aa673dfd5376