export `require this as` to hhvm and add trait constraint checks at class load time
Summary:
`require this as` trait constraints can be seen as `require extends` constraint without the *strict* subtype limitation. In other terms, if a trait `T` has a `require this as C` trait constraint, then the trait can be used by `C` and all subclasses of `C`.

This diff exports the `require this as` trait constraints to HHVM (`emit_class.rs`) and implements the trait-use checks at class load time (`class.cpp`).

Reviewed By: sinancepel

Differential Revision: D74886705

fbshipit-source-id: 50658f6c59d7ea3da255f298d5941d76c09eae02