Improve forwarding of safe_alias through coro functions & wrappers
Summary:
The theme here is "making `safe_alias` annotations pervasive, with minimal overhead". The immediate goal for this diff was for `collectAll()` and `timeout()` to return a `SafeTask` if the inputs are safe. This is quite helpful to up-stack adoption of `SafeTask`.

My actual motivation was that I needed the up-stack result-collector primitive of `co_result_transform` to have the same properties of "transparent to safe_alias, await-immediately, and noexcept-awaitable". This diff makes that trivial to implement.

The main changes:

(1) A type's `safe_alias_of` can be set via a specialization (as before), or via the member type `folly_private_safe_alias_t` (new). It is required since I cannot add specializations for member types -- those are non-deducible -- but e.g. `AsyncGenerator<...>::NextAwaitable` must be `safe_alias`-tagged in practice. The alias is also shorter, which is nice.

(2) Migrate `collectAll()` and `timeout()` to use `PickTaskWrapper`, which generalizes its predecessor `most_safe_task_type` (aka `best_fit_task_wrapper`) to allow the user to select all 3 major dimensions of task wrappers:
  * `safe_alias_of`
  * `must_await_immediately_v`
  * `is_noexcept_awaitable_v` (via the "add wrapper" feature) -- not yet used for `collectAll`, not applicable to `timeout`

(3) While adding `folly_private_safe_alias_t` support, I updated `SafeAlias.h` and `SafeAlias-fwd.h` for clarity & simplicity. There's also a minor simplification to `SafeTask.h` (and I have one more to do up-stack).

Note that, as a consequence of (2), I made both `Collect.h` and `Timeout.h` unconditionally depend on `SafeTask.h` and `NowTask.h`. This is not ideal from the point of view of compilation size, but it's the only non-confusing behavior, since both of those combinators can emit a different task type than their input (comments inline).

Reviewed By: ispeters

Differential Revision: D77704741

fbshipit-source-id: b347ea7830b99b71fb66b56da57a6774797e95b3