take plain string literals in exception_shared_string
Summary:
The high-level goal is to be able to use the "bit-variant" design of `exception_shared_string` to store either a refcounted heap string, or a string literal pointer in 8 bytes.

Prior to this change, this is only possible by passing a string-literal-tagged-type aka `"a"_litv`. This works robustly, but has very poor usability. E.g. in my error-stacking library for `folly::result` coroutines, end-users would have to know to write this:

```
using namespace folly::string_literals;
co_await with_stack_error(returnsResult(), "message"_litv);
```

Instead of the much more obvious:

```
co_await with_stack_error(returnsResult(), "message");
```

This change supports the  "natural" syntax by using the top pointer bit for the variant.

We can no longer use one of the pointer alignment bits for the variant since, unfortunately, string literals can be stored at both even and odd static addresses.  On the other hand, AFAIK, the top 4 bits of the 64 bits of a pointer are free on ARM and X64, even when memory tagging extensions are in use.

For 32-bit systems, the story of the top bit is murkier. On most platforms, userspace pointers are limited to 2GiB and the top bit is zero, but there appear to be some exceptions. `folly` is not supported on any of the "can use the top bit" 32-bit platforms, but to avoid silent corruption in the future, there's a production 3-bit-only `CHECK` against such pointers.

## Removing dubious `strlen` constructor

Besides the main change, I also removed the now-conflicting `strlen` constructor `exception_shared_string(const char*)`. IIUC, this was practically unused, and moreover is potentially harmful since it results in a buffer overrun bugs if someone inadvertently passes `std::string::data()` or similar.

## Design notes

Here's how I think about the design constraints for making the "natural syntax" work:
 - IIUC, in C++ it's impossible by-design for a consteval function call to allocate new static storage. The folks who designed `consteval` were strongly against making them like templates.
 - There's no `constexpr` parameter either to bridge from "function parameter" to "new type", and it's probably not coming -- e.g. p1045 was stuck since 2019.
 - This means that if we take `const char*` string literals, there is no control over their address at all.
 - Thus, we either need some OS-level way of distinguishing heap pointers from static pointers (seems brittle), or we need an unused bit (as is done here).
 - An non-bit-hack alternative would be to have `what()` load the first byte of the pointed-to data, and use: `what_ == nullptr` for the empty string, a NUL byte for allocated strings, a non-NUL byte for literal strings.

Reviewed By: ispeters

Differential Revision: D74196963

fbshipit-source-id: 480e4bb96f7fc80f630b9f0a037663277659b4a0