Improve completeness wrt like pushing
Summary:
Under implicit pessimisation, `hh` fails to infer types in situations that appear quite "easy". Examples are the new tests in this diff. The fundamental issue is that "like pushing" sometimes doesn't kick in if the type contains type variables, because Hack is assuming that `vec<#0>` (for example) does not support dynamic. But if `#0` has any bound that supports dynamic (as it will do always, under implicit pessimisation), then we should regard `vec<#0>` as supporting dynamic. To this end, we replace the existing code for `is_supportdyn` that attempts a subtype against `supportdyn<mixed>` with explicit recursion over the type structure, which looks at bounds on type variables.

Unfortunately this uncovered further issues, which are rectified in this diff: see comments on individual lines.

Differential Revision: D77376309

fbshipit-source-id: c11f3701d3b2fb5a0316b6468dbb981e609858c3