don't cache singleton defaults
Summary:
There shouldn't be much benefit to caching the default `int`, `bool`, `b""`, or `()` values. That's because these are singletons. "Creating" one of these skips the costly Python object allocation, so it may be cheaper to return the singleton directly than to search the map for it.

This also helps us net-remove one branch from this function.

Reviewed By: createdbysk, prakashgayasen

Differential Revision: D77685921

fbshipit-source-id: 31dcffdd925fdf7685c8b70c0963f977d36a4101