Enable ZDB mode based on features file
Summary:
# Context
We currently use both the features file and manually setting the ZDB mode
to indicate whether to use hybrid or no rich client client. The former is preferred as we
can enable/disable things on the fly.

# This Diff
Removed flag to set client mode which is currently set to hybrid but we
disabled hybrid in the features file to avoid going to the CAS server.

Reviewed By: nt591

Differential Revision: D77956726

fbshipit-source-id: 1e873f3287ec0521c8d79d94b76061cd864e8fcb