Mangle metadata import paths
Summary:
The solution to {D77494749} was to use the mangled import name.

The previous diff in this stack {D77617349} added yet another import of an external module for metadata.

To preempt more symbol conflicts that might occur in the future, change all metadata imports to use the mangled module names.

Reviewed By: yfeldblum

Differential Revision: D77617854

fbshipit-source-id: 37c84ed103950d544c3ec3d91e735323018f948f