cache more defaults
Summary:
Remove some complexity added by D77685921.
This was one of two strategies to reduce cost of this function. The other was switching the map type to `F14FastMap`, which did work (50% reduction). However, apparent gCPU cost of function went up about 20%, so apparently caching the basic unqualified types like `0`, `b""` etc. is better.

Even if the perf is a wash, better to reduce the SLOC and duplicated logic.

Reviewed By: prakashgayasen

Differential Revision: D78036928

fbshipit-source-id: f416bbec076bc86e03e6a2fd47a534b25c53119d