Add supportdyn<mixed> as an upperbound for the tyvar for the value generic of a collection literal
Summary:
I discovered this specific problem when investigating a more general problem with my class type refinement logic. (repro is the new test case "collection_generic_supportdyn_upper_bound.php")

From T228583128:

> When we do the intersection between the scrutinee type and the hint type during refinement, we do a trick to make the hint side of the intersection supportdyn if the scrutinee is supportdyn. In the above example, the scrutinee does not pass the "is supportdyn" check because it's a union with a tyvar with no upperbound.

> code pointer: https://fburl.com/code/aa55qq9v

> Later when we do the subtyping check for passing $item to takes_container the Container part of the intersection fails due to the supportdyn mismatch and the tyvar part of the intersection fails due to the tyvar having an incompatible lower bound.

I talked to andrewjkennedy about this and he suggested that Vector's generic tyvar should have an upper bound of `supportdyn<mixed>`, so this is what this diff does.

I don't think this is a general solve for the incompleteness I mention since the "is supportdyn" check we do during refinement is pretty bad with tyvars generally (really only if the tyvar has exactly the given supertype will the is_subtype check work), but if this change is correct in principal then, might as well.

I'm not super familiar with the collection typing code, so this could use a review by someone who feels more confident than me.

Reviewed By: andrewjkennedy

Differential Revision: D77973689

fbshipit-source-id: 1f0630f76ded89aec8b9a56b0fb29476e140b117