better integration with streaming errors file in client
Summary:
## Context

Prior to D77791536, `hh_client check` does the following when streaming errors are enabled:
It looks for an errors file produced by the server. If it finds one, it looks for a Watchman clock saved in the header of the errors file.
- If a Watchman clock `c` os present in the file, `hh_client` sends a request to Watchman to check if there have any file changes since `c`:
  + If yes, we don't use the errors file and wait for the server to notice the file changes, start a re-check and create a new errors file that we should be using instead.
  + Otherwise, we actually use the errors file.
- If there is no clock saved in the errors file, we use the errors file anyway. In the worst case, users will just see a message that the type-check was restarted due to changed files and they need to re-run `hh`.

D77791536 changed this so that we may now read a clock from the errors file that was created by `Edenfs_watcher` instead. In that case, we would handle it like the second case above ( = use it anyway without actually looking at the clock, tolerating the slightly worse UX)

## In this diff

This diff changes the logic outline above so that if the errors file contains a clock, we check if it's a Watchman or Eden clock, and ask the respective file watcher if there have been changes since that clock.

For `Edenfs_watcher`, this is implemented using the new function `Edenfs_watcher.Standalone.get_changes_since`. It is "standalone" in the sense that it does not create a `EdenfsWatcherInstance` that would get changes in the background.

Note that this overall design means that we automatically choose the right file watcher to ask in the client based on the file watcher that was used in the server to create the errors file.

## Caveat
Note that this diff requires changing `Folly.ensure_init` to `Startup_initializer.init` in the client. This means that we perform `fbinit` now instead of just initializing folly. This is required in order to use `Edenfs_watcher` in the client.

This has the downside of making `hh_client` print an annoying message to `stderr` in some situations. Details described here: T230278435

Differential Revision: D77935368

fbshipit-source-id: 00de19a03224994ac01e177b4a8578914ec0c93f